requirements.txt
# Your existing packages
playwright==1.55.0
pdfplumber
pandas
rapidfuzz
beautifulsoup4
tqdm
python-dateutil
streamlit
requests
sqlite-utils

# New packages for GPT-based parser
openai>=1.3.0
python-dotenv>=1.0.0
supabase>=2.0.0

